requirements.txt
# Data Processing
pandas==2.1.0
numpy==1.24.3
polars==0.19.0

# Database
duckdb==0.9.2
sqlalchemy==2.0.23

# Visualization & Dashboard
streamlit==1.28.0
plotly==5.17.0
altair==5.1.0

# dbt
dbt-core==1.7.0
dbt-duckdb==1.7.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
faker==20.0.0
loguru==0.7.2

# Development
jupyter==1.0.0
ipykernel==6.25.0
pytest==7.4.0

# Data Quality (simplified - we'll add later if needed)
# great-expectations==0.18.8  # Commented out for now
